+++ server/controllers/index.ts
@@ -11,6 +11,7 @@ import AllocationsController from './tasks/allocationsController'
 import type { Services } from '../services'
 import DashboardController from './dashboardController'
 import PagesController from './placementApplications/pagesController'
+import PeopleController from './people/peopleController'
 import ReviewController from './placementApplications/reviewController'
 import WithdrawalsController from './placementApplications/withdrawalsController'
 import RedirectController from './redirectController'
@@ -31,6 +32,7 @@ export const controllers = (services: Services) => {
   )
   const placementApplicationReviewController = new ReviewController(services.placementApplicationService)
   const placementApplicationWithdrawalsController = new WithdrawalsController(services.placementApplicationService)
+  const peopleController = new PeopleController(services.personService)
 
   return {
     redirectController,
@@ -40,6 +42,7 @@ export const controllers = (services: Services) => {
     placementApplicationPagesController,
     placementApplicationReviewController,
     placementApplicationWithdrawalsController,
+    peopleController,
     ...applyControllers(services),
     ...assessControllers(services),
     ...matchControllers(services),
