+++ server/controllers/people/peopleController.test.ts
@@ -2,10 +2,10 @@ import type { NextFunction, Request, Response } from 'express'
 import { DeepMocked, createMock } from '@golevelup/ts-jest'
 
 import PeopleController from './peopleController'
-import PersonService from '../services/personService'
-import { personFactory } from '../testutils/factories'
-import { errorMessage, errorSummary } from '../utils/validation'
-import { RestrictedPersonError } from '../utils/errors'
+import PersonService from '../../services/personService'
+import { personFactory } from '../../testutils/factories'
+import { errorMessage, errorSummary } from '../../utils/validation'
+import { RestrictedPersonError } from '../../utils/errors'
 
 describe('PeopleController', () => {
   const token = 'SOME_TOKEN'
