+++ server/routes/index.ts
@@ -13,6 +13,7 @@ import manageRoutes from './manage'
 import tasksRoutes from './tasks'
 import placementApplicationRoutes from './placementApplications'
 import adminRoutes from './admin'
+import peopleRoutes from './people'
 
 export default function routes(controllers: Controllers, services: Partial<Services>): Router {
   const router = Router()
@@ -30,6 +31,7 @@ export default function routes(controllers: Controllers, services: Partial<Servi
   tasksRoutes(controllers, router, services)
   placementApplicationRoutes(controllers, router, services)
   adminRoutes(controllers, router, services)
+  peopleRoutes(controllers, router, services)
 
   return router
 }
